requirements.txt
asgiref
Django
sqlparse
typing_extensions
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-phonenumber-field
phonenumbers
factory-boy
drf-spectacular
# Validation and Formaters -> delete for production
flake8
flake8-docstrings
pytest
pytest-django
pytest-cov
django-extensions
faker
pytz
django-storages
python-decouple
boto3
celery[redis]
# will be deleted in future versions
drf-writable-nested
redis
PyJWT
django-filter
drf-excel
psycopg2